Move quat to euler decomposition to math class.
Previously wanted it in float3 but that's problematic because Burst
can fail to compile a vector type if it contains methods it doesn't
expect.
