--- src/Unity.Mathematics/float3.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dee9c109b55974803bf35d5346b202cf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
